+++ external-files-from-google-drive.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * Plugin Name:       External files from Google Drive in Media Library
- * Description:       Extends External files in Media Library for usage of Google Drive.
+ * Description:       Extends External files in Media Library for use with Google Drive.
  * Requires at least: 6.2
  * Requires PHP:      8.1
  * Requires Plugins:  external-files-in-media-library