+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Updates translations
+
 ## [1.0.1] - 29.03.2026
 
 ### Added